+++ CSharpSem/Tasks/Task27.cs
@@ -9,13 +9,13 @@ internal class Task27 : Task, IRunnableFromConsole
     {
         Console.WriteLine(description);
 
-        long number = EdInput.InputConsoleInteger("Введите натуральное число", Algebra.Sets.N);
+        int number = EdInput.InputConsoleInteger("Введите натуральное число", Algebra.Sets.N);
 
         Console.WriteLine($"Сумма цифр числа {number} равна {s_sumOfArray(s_getDigits(number))}");
     }
 
     // Считает количество разрядов числа.
-    private static int s_getDepth(long number)
+    private static int s_getDepth(int number)
     {
         int depth = 0;
         while (number > 0)
@@ -27,12 +27,12 @@ internal class Task27 : Task, IRunnableFromConsole
     }
 
     // Создаёт массив из цифр числа.
-    private static int[] s_getDigits(long number)
+    private static int[] s_getDigits(int number)
     {
         var digits = new int[s_getDepth(number)];
         for (int i = 0; i < digits.Length; i++)
         {
-            int currentDigit = (int)number % 10;
+            int currentDigit = number % 10;
             number /= 10;
             digits[^(i + 1)] = currentDigit;
         }
